Put PermissionAttachmentManager in seperate file
